Refactor project structure and update packages
This commit does a major restructuring of the project and includes updates to packages. It features changes in the location for common properties and updates to the version of xunit. The publish-windows PowerShell script has been removed due to changes in the project structure. It also comes with modifications in the project files to accommodate the new structure and adjustments to the "README.md" file to explain the application installation and usage.