Refactor and simplify CLI installer actions
The commit primarily removes and refactors actions in the CLI installer in preparation for a more streamlined installation process. Major changes include the removal of multiple action-specific files and the addition of installation logic in relevant command files. Additionally, several helper classes have been refactored as well to aid in the application's installation processes.